+++ AdventOfCode/IntCodes/InstructionParser.cs
@@ -5,10 +5,10 @@ namespace AdventOfCode.IntCodes
     public static class InstructionParser
     {
         //turns a string into instruction and parameter array
-        public static int[] Parse(string instructions)
+        public static long[] Parse(string instructions)
         {
             string[] commands = instructions.Split(',');
-            return Array.ConvertAll(commands, int.Parse);
+            return Array.ConvertAll(commands, long.Parse);
         }
     }
 }
